+++ src/Plugins/SmartStore.PayPal/Providers/PayPalExpressProvider.cs
@@ -21,6 +21,7 @@ using SmartStore.Services.Directory;
 using SmartStore.Services.Orders;
 using SmartStore.Services.Payments;
 using SmartStore.Services.Shipping;
+using SmartStore.Web.Framework.Plugins;
 
 namespace SmartStore.PayPal
 {
@@ -38,7 +39,7 @@ namespace SmartStore.PayPal
         private readonly ICustomerService _customerService;
         private readonly ICountryService _countryService;
         private readonly HttpContextBase _httpContext;
-        
+
         public PayPalExpress(
             ICurrencyService currencyService,
             IPriceCalculationService priceCalculationService,
@@ -59,6 +60,7 @@ namespace SmartStore.PayPal
             _customerService = customerService;
             _countryService = countryService;
             _httpContext = httpContext;
+
         }
 
 		protected override string GetResourceRootKey()
@@ -76,7 +78,7 @@ namespace SmartStore.PayPal
 			var result = new ProcessPaymentResult();
             var doPayment = DoExpressCheckoutPayment(processPaymentRequest);
 			var settings = Services.Settings.LoadSetting<PayPalExpressPaymentSettings>(processPaymentRequest.StoreId);
-
+            
             if (doPayment.Ack == AckCodeType.Success)
             {
                 if (PayPalHelper.GetPaymentAction(settings) == PaymentActionCodeType.Authorization)
@@ -121,7 +123,16 @@ namespace SmartStore.PayPal
             // build the request
             var req = new DoCaptureReq
             {
-                DoCaptureRequest = new DoCaptureRequestType()
+                DoCaptureRequest = new DoCaptureRequestType
+                {
+                    Version = PayPalHelper.GetApiVersion(),
+                    AuthorizationID = capturePaymentRequest.Order.CaptureTransactionId,
+                    Amount = new BasicAmountType { 
+                        Value = Math.Round(capturePaymentRequest.Order.OrderTotal, 2).ToString("N", new CultureInfo("en-us")),
+                        currencyID = (CurrencyCodeType)Enum.Parse(typeof(CurrencyCodeType), Helper.CurrencyCode, true)
+                    },
+                    CompleteType = CompleteCodeType.NotComplete
+                }
             };
 
             //execute request
@@ -138,6 +149,7 @@ namespace SmartStore.PayPal
             {
                 capturePaymentResult.CaptureTransactionId = result.DoCaptureResponseDetails.PaymentInfo.TransactionID;
                 capturePaymentResult.CaptureTransactionResult = "Success";
+                capturePaymentResult.NewPaymentStatus = PaymentStatus.Paid;
             }
             else
             {